refactor: 都道府県関連のデータを取得する関数が population ディレクトリにあったため、utils/prefecture に移動
